Split featured asset components into separate files
